refector file structs
