+++ celerybeatredis/exceptions.py
@@ -1,0 +1,9 @@
+# Copyright 2014 Kong Luoxing
+
+# Licensed under the Apache License, Version 2.0 (the 'License'); you may not
+# use this file except in compliance with the License. You may obtain a copy
+# of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+
+class ValidationError(Exception):
+    pass
